+++ src/Core/Backend.Core.DDD/Queries/IQueryBus.cs
@@ -1,9 +1,10 @@
-﻿using System.Threading.Tasks;
+﻿using System.Threading;
+using System.Threading.Tasks;
 
 namespace GoldenEye.Backend.Core.DDD.Queries
 {
     public interface IQueryBus
     {
-        Task<TResponse> Send<TQuery, TResponse>(TQuery query) where TQuery : IQuery<TResponse>;
+        Task<TResponse> SendAsync<TQuery, TResponse>(TQuery query, CancellationToken cancellationToken = default(CancellationToken)) where TQuery : IQuery<TResponse>;
     }
 }