+++ src/Core/Shared.Core/Modules/Module.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace GoldenEye.Shared.Core.Modules
+{
+    public abstract class Module : IModule
+    {
+        public virtual void Configure(IServiceCollection services)
+        {
+        }
+
+        public virtual void Use()
+        {
+        }
+    }
+}