+++ src/Core/Backend.Core.WebApi.Tests/Changelog.md
@@ -1,4 +1,10 @@
-﻿# v2.3.0 (29.12.2017)
+﻿# v3.0.9 (18.04.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/57)
+
+## Changes
+
+* updated packages version to most recent **[PATCH]**
+
+# v2.3.0 (29.12.2017)
 
 ## Changes