+++ src/Sample/DDD/Backend.DDD.Sample.IntegrationTests/Changelog.md
@@ -1,4 +1,17 @@
-﻿# v3.0.7 (07.04.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/53)
+﻿# v4.0.1 (18.04.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/57)
+
+## Changes
+
+* updated packages version to most recent **[PATCH]**
+
+# v4.0.0 (18.04.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/56)
+
+## Changes
+
+* Added Async sufix for Publish, Send methods of CommandBus, QueryBus, EventBus abd proper cancellation token handling **[MAJOR]**
+
+
+# v3.0.7 (07.04.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/53)
 
 ## Changes